'merge two rasters'

+++ tools/combine_rasters_and_vectorization.py
@@ -121,8 +121,8 @@ def vectorize_image(raster_1, raster_2, output_file, category, threshold):
     im2 = rasterio.open(raster_2)
     data_2 = im2.read(1)
 
-    data_1 = data_1 / 2.0
-    data_2 = data_2 / 2.0
+    data_1 = np.int(data_1 / 2)
+    data_2 = np.int(data_2 / 2)
 
     data = data_1 + data_2
